requirements.txt
# Database dependencies
# neo4j==5.14.1
# chromadb==0.4.22

# AI and embeddings (COMMENTED - uncomment when needed)
# torch==2.5.1+cu121
# transformers==4.36.2
# sentence-transformers==2.2.2
# llama-cpp-python==0.1.85
# langchain==0.0.335
# langchain-community==0.0.16
# numpy==1.26.4
# tensorflow==2.18.0

# Document processing (COMMENTED - uncomment when needed)
# pypdf2==3.0.1
# pdfplumber==0.10.3
# pdf2image==1.17.0
# tiktoken==0.5.2

# ===================================
# WEB FRAMEWORK & API
# ===================================
fastapi==0.115.14
uvicorn[standard]==0.35.0
gunicorn==22.0.0
pydantic==2.10.4
pydantic-settings==2.8.0

# Real-time WebSocket communication for React frontend
python-socketio==5.11.4
python-engineio==4.11.1

# File uploads and form handling
python-multipart==0.0.20
aiofiles==24.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Template engine
jinja2==3.1.2

# ===================================
# UTILITIES
# ===================================
loguru==0.7.2
tqdm==4.66.1
pandas==2.1.3
httpx==0.28.1
requests==2.31.0
python-dotenv==1.0.0

# ===================================
# JUPYTER & PROJECT MANAGEMENT
# ===================================
jupyter>=1.0.0
notebook>=6.5.4
ipykernel>=6.22.0
nbconvert>=7.3.1
ipywidgets==8.1.1

# ===================================
# TESTING (FIXED VERSIONS)
# ===================================
pytest>=8.2.0
pytest-cov>=4.1.0
pytest-asyncio>=0.24.0

# ===================================
# DEVELOPMENT TOOLS
# ===================================
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# ===================================
# DATABASE - PostgreSQL for Sermon Organizer
# ===================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
sqlmodel==0.0.14
asyncpg==0.29.0